fix: rename llmwiki/queue.py → ingest_queue.py to unshadow stdlib (#491, v1.3.4)
Naming a module 'queue.py' shadows Python's stdlib 'queue' module,
breaking any future code inside llmwiki/ wanting queue.Queue for
thread-safe primitives. Pylint/ruff also flag this anti-pattern.

Renamed the module to llmwiki/ingest_queue.py (matches the actual
purpose — pending-source ingest queue, not a generic queue).

Old llmwiki/queue.py becomes a back-compat shim that re-exports
the public API and emits a DeprecationWarning. Will be removed
in v1.5 after one minor cycle.

Tests updated:
- test_queue.py imports from llmwiki.ingest_queue
- new test_ingest_queue_shim.py covers (1) the canonical module
  imports cleanly, (2) the legacy shim still works + warns, (3)
  bare 'import queue' from inside llmwiki/ now resolves to stdlib
  (the reason for the rename)
