+++ tests/test_queue.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from llmwiki.queue import enqueue, dequeue, peek, clear, queue_size
+from llmwiki.ingest_queue import enqueue, dequeue, peek, clear, queue_size
 
 
 # ─── Basic operations ──────────────────────────────────────────────────
